DataElementRoot extends new AnnotatedClass
This provides us a good starting point for later adding DataModel
